+++ .eslintrc.json
@@ -187,7 +187,6 @@
         "no-new-wrappers": "error",
         "no-nonoctal-decimal-escape": "error",
         "no-octal-escape": "error",
-        "no-param-reassign": "error",
         "no-plusplus": "off",
         "no-promise-executor-return": "error",
         "no-proto": "error",
