+++ comandos/utilitarios/ajuda.js
@@ -1,5 +1,6 @@
 const { MessageButton, MessageEmbed } = require("discord.js");
-const { formatarCanal, traduzirPerms } = require("../../modulos/utils");
+const { traduzirPerms } = require("../../modulos/utils");
+const { pagina } = require("../../modulos/interações")
 
 module.exports = {
     //* Infomações do comando
@@ -89,18 +90,18 @@ module.exports = {
             }
             embedsarray.unshift(menuEmbed.setFooter(`Veja outras páginas, clicando nos botões • Página 0/${embedsarray.length}`))
 
-            var pagina = 0;
+            var paginaAtual = 0;
 
             const voltar = new MessageButton()
                 .setCustomId(`voltar`)
                 .setLabel('<<')
-                .setDisabled(false)
+                .setDisabled(true)
                 .setStyle('SECONDARY');
 
             const menu = new MessageButton()
                 .setCustomId('menu')
                 .setLabel('O')
-                .setDisabled(false)
+                .setDisabled(true)
                 .setStyle("PRIMARY")
 
             const progredir = new MessageButton()
@@ -112,116 +113,64 @@ module.exports = {
             const resposta = await msg.channel.send({
                 content: null,
                 embeds: [menuEmbed],
-                components: [[
-                    voltar.setDisabled(true),
-                    menu.setDisabled(true),
-                    progredir
-                ]],
+                components: [[voltar, menu, progredir]],
                 reply: { messageReference: msg }
             }).catch();
 
             const respostas = {
                 voltar(i) {
-                    if (pagina === 0) return client.log("aviso", `Comando "${module.exports.nome}" com paginas dessincronizadas (${msg.id})`);
+                    if (paginaAtual === 0) return client.log("aviso", `Comando "${module.exports.nome}" com paginas dessincronizadas (${msg.id})`);
 
-                    --pagina
+                    --paginaAtual
                     i.update({
                         content: resposta.content || null,
-                        embeds: [embedsarray[pagina].setFooter(`Veja outras páginas, clicando nos botões • Página ${pagina}/${embedsarray.length - 1}`)],
+                        embeds: [embedsarray[paginaAtual].setFooter(`Veja outras páginas, clicando nos botões • Página ${paginaAtual}/${embedsarray.length - 1}`)],
                         components: [[
-                            voltar.setDisabled(pagina <= 0),
-                            menu.setDisabled(pagina <= 0),
-                            progredir.setDisabled(embedsarray.length - 1 <= pagina)
+                            voltar.setDisabled(paginaAtual <= 0),
+                            menu.setDisabled(paginaAtual <= 0),
+                            progredir.setDisabled(embedsarray.length - 1 <= paginaAtual)
                         ]]
                     }).catch();
+
+                    return { parar: false, paginaAtual, paginaTotal: embedsarray.length - 1 }
                 },
                 menu(i) {
-                    if (pagina === 0) return client.log("aviso", `Comando "${module.exports.nome}" com paginas dessincronizadas (${msg.id})`);
+                    if (paginaAtual === 0) return client.log("aviso", `Comando "${module.exports.nome}" com paginas dessincronizadas (${msg.id})`);
 
-                    pagina = 0
+                    paginaAtual = 0
                     i.update({
                         content: resposta.content || null,
-                        embeds: [embedsarray[pagina].setFooter(`Veja outras páginas, clicando nos botões • Página ${pagina}/${embedsarray.length - 1}`)],
+                        embeds: [embedsarray[paginaAtual].setFooter(`Veja outras páginas, clicando nos botões • Página ${paginaAtual}/${embedsarray.length - 1}`)],
                         components: [[
-                            voltar.setDisabled(pagina <= 0),
-                            menu.setDisabled(pagina <= 0),
-                            progredir.setDisabled(embedsarray.length - 1 <= pagina)
+                            voltar.setDisabled(paginaAtual <= 0),
+                            menu.setDisabled(paginaAtual <= 0),
+                            progredir.setDisabled(embedsarray.length - 1 <= paginaAtual)
                         ]]
                     }).catch();
+
+                    return { parar: false, paginaAtual, paginaTotal: embedsarray.length - 1 }
                 },
                 progredir(i) {
-                    if (embedsarray.length - 1 <= pagina) return client.log("aviso", `Comando "${module.exports.nome}" com paginas dessincronizadas (${msg.id})`);
+                    if (embedsarray.length - 1 <= paginaAtual) return client.log("aviso", `Comando "${module.exports.nome}" com paginas dessincronizadas (${msg.id})`);
 
-                    ++pagina
+                    ++paginaAtual
                     i.update({
                         content: resposta.content || null,
-                        embeds: [embedsarray[pagina].setFooter(`Veja outras páginas, clicando nos botões • Página ${pagina}/${embedsarray.length - 1}`)],
+                        embeds: [embedsarray[paginaAtual].setFooter(`Veja outras páginas, clicando nos botões • Página ${paginaAtual}/${embedsarray.length - 1}`)],
                         components: [[
-                            voltar.setDisabled(pagina <= 0),
-                            menu.setDisabled(pagina <= 0),
-                            progredir.setDisabled(embedsarray.length - 1 <= pagina)
+                            voltar.setDisabled(paginaAtual <= 0),
+                            menu.setDisabled(paginaAtual <= 0),
+                            progredir.setDisabled(embedsarray.length - 1 <= paginaAtual)
                         ]]
                     }).catch();
+
+                    return { parar: false, paginaAtual, paginaTotal: embedsarray.length - 1 }
                 }
             }
-            const coletor = resposta.createMessageComponentCollector({ time: 180000, idle: 60000 })
-            client.log("info", `Coletor de botões iniciado em #${formatarCanal(msg.channel)} por @${msg.author.tag} id:${msg.id}`)
 
-            coletor.on("collect", i => {
-                try {
-                    if (i.user.id !== msg.author.id) {
-                        const cuidaEmbed = new MessageEmbed()
-                            .setColor(client.defs.corEmbed.nao)
-                            .setTitle(`⛔ Cuida da sua vida`)
-                            .setDescription("essa mensagem não foi direcionada a você");
-                        return i.reply({ content: null, embeds: [cuidaEmbed], ephemeral: true })
-                    }
-
-                    respostas[i.customId](i);
-                    client.log("verbose", `@${i.user.tag} apertou "${i.customId}" pagina: ${pagina}/${embedsarray.length - 1} id:${msg.id}`)
-                } catch (err) {
-                    client.log("erro", err.stack)
-                    client.log("comando", `Ocorreu um erro em ${this.nome} ao ser executado por @${msg.author.tag}`, "erro");
-
-                    coletor.stop("erro")
-                }
-            })
-
-            coletor.once('end', (coletado, razao) => {
-                client.log("info", `Coletor de botões terminado por ${razao} em #${formatarCanal(msg.channel)}, coletando ${coletado.size} interações id:${msg.id}`);
-
-                let botoes = []
-                if (razao === "erro") {
-                    const erro = new MessageButton()
-                        .setCustomId(`erro`)
-                        .setLabel('Ocorreu um erro')
-                        .setDisabled(true)
-                        .setStyle('DANGER');
-                    botoes = [[erro]]
-                }
-                if (razao === "time") {
-                    const tempo = new MessageButton()
-                        .setCustomId(`tempo`)
-                        .setLabel("Tempo esgotado")
-                        .setDisabled(true)
-                        .setStyle('SECONDARY');
-                    botoes = [[tempo]];
-                }
-                if (razao === "idle") {
-                    const idle = new MessageButton()
-                        .setCustomId(`tempo`)
-                        .setLabel("Inatividade")
-                        .setDisabled(true)
-                        .setStyle('SECONDARY');
-                    botoes = [[idle]];
-                }
-
-                resposta.edit({
-                    content: resposta.content || null,
-                    embeds: resposta.embeds,
-                    components: botoes
-                }).catch();
-            });
+            //* Coletor de interações
+            const filtro = (i) => i.user.id !== msg.author.id
+            pagina(this, msg, resposta, respostas, filtro);
         }
     }
 };
